Splited class Song
Splitted to Classes Song and MidiFile_ and UltraStarFile, now the
format-specific code is in this classes. Song calls them for writing and
reading  over self.writer and self.reader.

Also tryed to track all dependencies to this class and updated them

besides the testing method now compares directly the songlist not the
md5sums of two pickle-dumps